+++ docs/CHANGELOG.md
@@ -58,6 +58,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove "is_major_incident" field from Flaw (OSIDB-2741)
 - Remove "meta_attr" field from FlawReference (OSIDB-2854)
 - Remove "meta_attr" field from FlawAcknowledgment (OSIDB-2854)
+- Remove "component" field from Flaw (OSIDB-2839)
 
 ## [3.7.3] - 2024-05-28
 ### Fixed